requirements.txt
streamlit>=1.32.0
--extra-index-url https://download.pytorch.org/whl/nightly/cu128
torch>=2.2.0
transformers>=4.38.0
diffusers>=0.25.0
accelerate>=0.27.0
bitsandbytes>=0.43.0
pillow>=10.0.0
pypdf2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
scipy>=1.11.0
safetensors>=0.4.0
sentencepiece>=0.1.99
transformers-stream-generator>=0.0.4
